refactor: 迁移超出职责范围的代码 (#872)
## 迁移内容

将以下文件从 execution/ 迁移到正确的模块：

1. **flow_dispatch.py** → orchestra/
   - Flow 管理逻辑，应属于编排层
   - 更新 10 个源文件 + 3 个测试文件的导入

2. **actor_support.py** → services/
   - Actor 支持服务，应属于服务层
   - 更新 4 个源文件 + 1 个测试文件的导入

3. **role_policy.py** → config/
   - 角色策略配置，应属于配置层
   - 更新 2 个源文件的导入

4. **store_context.py** → clients/
   - Store 上下文客户端，应属于客户端层
   - 更新 4 个源文件 + 1 个测试文件的导入

## 验证结果

- ✅ 所有测试通过（1682 passed, 1 skipped）
- ✅ mypy 类型检查通过（292 source files）
- ✅ 导入路径全部更新完成

## 影响范围

- 迁移文件：4 个
- 修改文件：23 个源文件 + 4 个测试文件
- 总计：541 行代码迁移到正确模块

Closes #872
