+++ src/vibe3/domain/handlers/issue_state_dispatch.py
@@ -4,10 +4,10 @@ import asyncio
 
 from loguru import logger
 
+from vibe3.clients.store_context import get_store
 from vibe3.config.orchestra_settings import load_orchestra_config
 from vibe3.domain.events.flow_lifecycle import ManagerDispatchIntent
 from vibe3.domain.handler_registry import register_handler
-from vibe3.execution.store_context import get_store
 from vibe3.models.orchestration import IssueInfo, IssueState
 from vibe3.roles.manager import build_manager_request
 from vibe3.services.issue_failure_service import block_manager_noop_issue
