+++ src/vibe3/roles/manager.py
@@ -13,7 +13,6 @@ from vibe3.config.settings import VibeConfig
 from vibe3.environment.session_naming import get_manager_session_name
 from vibe3.environment.session_registry import SessionRegistryService
 from vibe3.execution.contracts import ExecutionRequest
-from vibe3.execution.flow_dispatch import FlowManager
 from vibe3.execution.issue_role_support import (
     build_issue_async_cli_request,
     build_issue_sync_prompt_request,
@@ -23,6 +22,7 @@ from vibe3.execution.issue_role_support import (
 from vibe3.execution.role_contracts import MANAGER_GATE_CONFIG
 from vibe3.models.orchestra_config import OrchestraConfig
 from vibe3.models.orchestration import IssueInfo, IssueState
+from vibe3.orchestra.flow_dispatch import FlowManager
 from vibe3.prompts.manifest import PromptManifest, PromptProvider
 from vibe3.prompts.template_loader import _resolve_prompts_path
 from vibe3.roles.definitions import IssueRoleSyncSpec, TriggerableRoleDefinition
