+++ src/vibe3/services/handoff_service.py
@@ -9,11 +9,11 @@ from loguru import logger
 from vibe3.clients import SQLiteClient
 from vibe3.clients.git_client import GitClient
 from vibe3.exceptions import UserError
-from vibe3.execution.actor_support import (
-    extract_role_from_actor,
-)
 from vibe3.models.flow import FlowEvent
 from vibe3.models.verdict import VerdictRecord
+from vibe3.services.actor_support import (
+    extract_role_from_actor,
+)
 from vibe3.services.artifact_parser import ArtifactParser
 from vibe3.services.external_events import ExternalEventRecorder
 from vibe3.services.handoff_storage import HandoffStorage
