+++ tests/vibe3/domain/handlers/test_governance_scan.py
@@ -14,7 +14,7 @@ class TestGovernanceScanHandler:
     @patch("vibe3.agents.backends.codeagent.CodeagentBackend")
     @patch("vibe3.clients.sqlite_client.SQLiteClient")
     @patch("vibe3.services.orchestra_status_service.OrchestraStatusService")
-    @patch("vibe3.execution.flow_dispatch.FlowManager")
+    @patch("vibe3.orchestra.flow_dispatch.FlowManager")
     @patch("vibe3.execution.coordinator.ExecutionCoordinator")
     @patch("vibe3.domain.handlers.governance_scan.load_orchestra_config")
     def test_skips_when_governance_already_running(
@@ -54,7 +54,7 @@ class TestGovernanceScanHandler:
     @patch("vibe3.environment.session_registry.SessionRegistryService")
     @patch("vibe3.clients.sqlite_client.SQLiteClient")
     @patch("vibe3.services.orchestra_status_service.OrchestraStatusService")
-    @patch("vibe3.execution.flow_dispatch.FlowManager")
+    @patch("vibe3.orchestra.flow_dispatch.FlowManager")
     @patch("vibe3.execution.coordinator.ExecutionCoordinator")
     @patch("vibe3.orchestra.logging.append_governance_event")
     @patch("vibe3.domain.handlers.governance_scan.load_orchestra_config")
@@ -107,7 +107,7 @@ class TestGovernanceScanHandler:
     @patch("vibe3.environment.session_registry.SessionRegistryService")
     @patch("vibe3.clients.sqlite_client.SQLiteClient")
     @patch("vibe3.services.orchestra_status_service.OrchestraStatusService")
-    @patch("vibe3.execution.flow_dispatch.FlowManager")
+    @patch("vibe3.orchestra.flow_dispatch.FlowManager")
     @patch("vibe3.orchestra.logging.append_governance_event")
     @patch("vibe3.domain.handlers.governance_scan.load_orchestra_config")
     def test_no_dispatch_when_circuit_breaker_open(
@@ -143,7 +143,7 @@ class TestGovernanceScanHandler:
     @patch("vibe3.environment.session_registry.SessionRegistryService")
     @patch("vibe3.clients.sqlite_client.SQLiteClient")
     @patch("vibe3.services.orchestra_status_service.OrchestraStatusService")
-    @patch("vibe3.execution.flow_dispatch.FlowManager")
+    @patch("vibe3.orchestra.flow_dispatch.FlowManager")
     @patch("vibe3.execution.coordinator.ExecutionCoordinator")
     @patch("vibe3.orchestra.logging.append_governance_event")
     @patch("vibe3.domain.handlers.governance_scan.load_orchestra_config")
@@ -187,7 +187,7 @@ class TestGovernanceScanHandler:
     @patch("vibe3.environment.session_registry.SessionRegistryService")
     @patch("vibe3.clients.sqlite_client.SQLiteClient")
     @patch("vibe3.services.orchestra_status_service.OrchestraStatusService")
-    @patch("vibe3.execution.flow_dispatch.FlowManager")
+    @patch("vibe3.orchestra.flow_dispatch.FlowManager")
     @patch("vibe3.execution.coordinator.ExecutionCoordinator")
     @patch("vibe3.orchestra.logging.append_governance_event")
     @patch("vibe3.domain.handlers.governance_scan.load_orchestra_config")
