+++ tests/vibe3/services/test_flow_lifecycle.py
@@ -113,8 +113,8 @@ def test_reactivate_flow_preserves_initiator():
 
 def test_flow_manager_uses_service_for_reactivation():
     """FlowManager should delegate to FlowService for reactivation."""
-    from vibe3.execution.flow_dispatch import FlowManager
     from vibe3.models.orchestra_config import OrchestraConfig
+    from vibe3.orchestra.flow_dispatch import FlowManager
 
     config = OrchestraConfig(repo="test/repo")
     manager = FlowManager(config)
