+++ tests/vibe3/test_verdict_service.py
@@ -5,8 +5,8 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from vibe3.execution.actor_support import extract_role_from_actor
 from vibe3.models.verdict import VerdictRecord
+from vibe3.services.actor_support import extract_role_from_actor
 from vibe3.services.verdict_service import VerdictService
